project corrrections

+++ 16-binary_tree_is_perfect.c
@@ -9,7 +9,7 @@
  */
 int binary_tree_is_perfect(const binary_tree_t *tree)
 {
-	size_t left_height, right_height;
+	int left_height, right_height;
 
 	if (tree == NULL)
 		return (0);
